Remove CallbackStore
