+++ src/Repository/CallbackRepository.php
@@ -40,4 +40,24 @@ class CallbackRepository extends ServiceEntityRepository
     {
         return $this->count(['type' => $type]) > 0;
     }
+
+    public function getFinishedCount(): int
+    {
+        return $this->count([
+            'state' => [
+                CallbackInterface::STATE_FAILED,
+                CallbackInterface::STATE_COMPLETE,
+            ],
+        ]);
+    }
+
+    /**
+     * @param CallbackInterface::TYPE_* $type
+     */
+    public function getTypeCount(string $type): int
+    {
+        return $this->count([
+            'type' => $type,
+        ]);
+    }
 }
