+++ src/Services/CallbackState.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\Services;
 
 use App\Repository\CallbackRepository;
-use App\Services\EntityStore\CallbackStore;
 
 class CallbackState implements \Stringable
 {
@@ -14,7 +13,6 @@ class CallbackState implements \Stringable
     public const STATE_COMPLETE = 'complete';
 
     public function __construct(
-        private CallbackStore $callbackStore,
         private CallbackRepository $repository
     ) {
     }
@@ -25,7 +23,7 @@ class CallbackState implements \Stringable
     public function __toString(): string
     {
         $callbackCount = $this->repository->count([]);
-        $finishedCallbackCount = $this->callbackStore->getFinishedCount();
+        $finishedCallbackCount = $this->repository->getFinishedCount();
 
         if (0 === $callbackCount) {
             return self::STATE_AWAITING;
